Fix corrupted content due to previously unresolved conflict
Ideally this should have been resolved before pushing to origin. However, I accidentally did push with the conflict, and as much as I would like to amend the commit to have a clean history, I won't in order to avoid causing problems for those who may have cloned the repo in the meantime.